[web] introduce confirmation alert component
Summary:
To confirm the deletion of a keyserver, I will need a very similar modal to the delete role modal. Rather than copying and pasting, I introduced a `ConfirmationAlert` component that can be used in both instances.

Depends on D10080

Test Plan: Confirmed that the delete role modal looked and behaved as before

Reviewers: rohan, inka, michal

Reviewed By: rohan

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D10081