[keyserver/lib/native] Introduce VOICED_IN_ANNOUNCEMENT_CHANNELS thread permission
Summary:
This is an intermediate step in being able to assign new members added to an announcement subchannel either the `Voiced` role or the `Members` role. For now, this thread permission only gives permission / removes permission to speak in a `COMMUNITY_ANNOUNCEMENT_ROOT`, and the next diff should cover using this for announcement subchannels.

This is the first part to [[ https://linear.app/comm/issue/ENG-5637/assign-voiced-or-default-role-when-adding-members-based-on-their | ENG-5637 ]]

Depends on D9804

Test Plan: Created a new community and confirmed that the `voiced_in_announcement_channels` was set for `Admins`. When toggling it for `Members` and new roles, I also confirmed that the `voiced_in_announcement_channels` got added to the role permission blobs. Lastly, I made sure that toggling this permission gave/removed the ability for users to speak in a COMMUNTIY_ANNOUNCEMENT_ROOT.

Reviewers: atul!, ginsu

Subscribers: ashoat, tomek, wyilio

Differential Revision: https://phab.comm.dev/D9827