[web/keyserver] make InitialReduxStateResponder configurable
Summary:
The purpose is to start rely on data in DB.
This change is allow endpoint to return only needed data, avoid fetching data persisted in DB.

This is safe for old clients, as this endpoint is only called after refresh/initial load, and those two things are equal with getting the newest version of the webapp.

This is only about changing endpoint, actual excluding data is implemented later in the stack.

Depends on D9948

Test Plan: Check if app still loads after refresh

Reviewers: tomek, michal

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D9949