[native] Introduce Rust encryption functions
Summary:
[ENG-5343](https://linear.app/comm/issue/ENG-5343)

Rust part of the AES encryption functions. For now the platform-specific code is temporary and will be filled in the next diffs with platform-specific encryption APIs.

Test Plan:
Run this code on iOS and Android:
```
fn test(promise_id: u32) {
  RUNTIME.spawn(async move {
    let f = move || -> Result<String, cxx::Exception> {
      let key = &mut [0; constants::aes::KEY_SIZE];
      ffi::generate_key(key)?;
      let plaintext = &mut [1, 2, 3, 4, 5, 6, 7, 8, 9, 10];
      let sealed_data =
        &mut [0; 10 + constants::aes::IV_LENGTH + constants::aes::TAG_LENGTH];
      ffi::encrypt(key, plaintext, sealed_data)?;
      let plaintext2 = &mut [0; 10];
      ffi::decrypt(key, sealed_data, plaintext2)?;

      Ok(format!(
        "
    Key: {key:?}
    Plaintext: {plaintext:?}
    Sealed Data: {sealed_data:?}
    Plaintext2: {plaintext2:?}"
      ))
    };
    handle_string_result_as_callback(f(), promise_id);
  });
}
```

Got the expected answer - the values didn't change (there is no logic yet, this is mostly to check for build issues).

Also modified `AESCrypto::encrypt` function to return a non-empty string (an error). This was correctly converted to an exception in C++ which was then converted to `Result::Err` in Rust.

Reviewers: marcin, varun, kamil

Reviewed By: marcin

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D9932