[keyserver] Block creating invite links for which an entry in Blob service already exists
Summary:
If a keyserver tries to create a link for which an entry exists, it means that a link was already created by some keyserver. Only one link can exist with a given name, so we have to block such an attempt.

Depends on D9943

Test Plan: Create a blob with `invite_${secret}` hash and check if it is impossible to create a link with the same secret. Also check if it is still possible to create other links.

Reviewers: michal, bartek, rohan

Reviewed By: rohan

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D9944