[web] start persisting `updatesCurrentAsOf`
Summary:
Some context is here: [ENG-5316](https://linear.app/comm/issue/ENG-5316/start-persisting-data-on-web-database#comment-767efe81).

We want to use the proper timestamp to allow the app to receive updates for data that will be stored in SQLite and longer received on initial load using `getInitialReduxStateResponder`.

This is a breaking change that could cause some issues, so launching this only for staff for now, putting this function in lib because later in the stack it is used also on a keyserver.

Test Plan:
1. For non-staff users make sure `updatesCurrentAsOf` is updated on each reload
2. For staff users, make sure it's persisted and updated after receiving an update
3. Check if the policies mechanism still works (it was connected, we used this to fetch all data after acknowledgment).

Reviewers: tomek, michal, inka

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D9948