[Flow202][native][skip-ci] [25/x] Get rid of ProfileScreen.navigateIfActive
Summary:
The new Flow wants us to type the parameter of this function, but it's actually rather annoying because the function only works for routes that don't have required params. To type it correctly, we'd have to enumerate all of the routes that the function needs to support.

Instead, this diff deletes this simple function in favor of calling `navigate` directly at callsites.

NOTE: CI will fail on this diff. I considered the possibility of fixing Flow errors BEFORE upgrading Flow, but it wasn't possible... in some cases, the fixes to support the new version of Flow caused errors in the old version. I could have hidden these type errors with `$FlowFixMe` lines and then later revert those, but that seemed like too much busy work.

Depends on D10008

Test Plan: Confirm the Flow errors go away

Reviewers: ginsu

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D10009