[keyserver] Update usage of threads.default_role in role-fetchers.js
Summary:
We need to get rid of usages of `threads.default_role` throughout the codebase so we can eventually drop the column.

This diff handles the usage in `role-fetchers.js`

Existing code: We determine whether or not a role is a default role based on `r.id = t.default_role` to then return in an object.

Depends on D9975

Addresses part of [[ https://linear.app/comm/issue/ENG-5833/update-code-that-uses-default-role-from-mariadb | ENG-5833 ]]

Test Plan: Ran the query before and after in my MariaDB on my local database and saw the same results printed out

Reviewers: atul, ginsu, tomek, ashoat

Reviewed By: ashoat

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D9976