[identity] Move GetOutboundKeysForUser to auth proto
Summary:
Copied the GetOutboundKeysForUser RPC from the client proto to the auth proto.
Kept references to the original request and response messages because they're incompatible with existing ones in the auth proto.

This one is going to be further refactored as a part of [[ https://linear.app/comm/issue/ENG-5659/use-user-id-for-lookup-instead-of-usernamewallet-address-in-x3dh | ENG-5659 ]].
Currently, left client side unchanged until that task is done.

Depends on D9931

Test Plan: Copy-paste so ensured Identity is building. The RPC and clients will be tested when refactored to use user ID instead of username/wallet address.

Reviewers: varun!

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D9957