[Flow202][native][skip-ci] [36/x] Type Redux state when it includes navState from dev tools
Summary:
We have a unique integration with `redux-devtools` that shows updates to ReactNav state and Redux state simultaneously. This diff introduces the type `MonitorActionState` for this combined state.

In addition, we have some custom code to support the dev tools user attempting to revert the state to a particular point. In those scenarios, `navState` actually gets inserted into the Redux state, and `NavFromReduxHandler` is responsible for extracting the nav state from the Redux state and putting it into ReactNav. This diff fixes up the types for `NavFromReduxHandler` as well.

NOTE: CI will fail on this diff. I considered the possibility of fixing Flow errors BEFORE upgrading Flow, but it wasn't possible... in some cases, the fixes to support the new version of Flow caused errors in the old version. I could have hidden these type errors with `$FlowFixMe` lines and then later revert those, but that seemed like too much busy work.

Depends on D10019

Test Plan: Confirm the Flow errors go away

Reviewers: atul

Reviewed By: atul

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D10020