[keyserver] Update usage of threads.default_role in validate-role-permissions.js
Summary: We need to get rid of usages of `threads.default_role` throughout the codebase so we can eventually drop the column.

This diff handles the usage in `validate-role-permissions.js`

Depends on D9977

Addresses part of [[ https://linear.app/comm/issue/ENG-5833/update-code-that-uses-default-role-from-mariadb | ENG-5833 ]]

Test Plan: Ran the script before and after the changes and logged the results of the query. I put them into a side by side text comparison on https://text-compare.com/, and confirmed that whenever default_role was equal to the roleID, the new is_default field = 1. Otherwise, it is null (casted to 0 during the `Boolean` conversion).

Reviewers: atul, ginsu, tomek

Subscribers: