[keyserver] Update usage of threads.default_role in thread-permission-updaters.js
Summary:
We need to get rid of usages of `threads.default_role` throughout the codebase so we can eventually drop the column.

This diff handles the usage in `thread-permission-updaters.js`

Depends on D9978

Addresses part of [[ https://linear.app/comm/issue/ENG-5833/update-code-that-uses-default-role-from-mariadb | ENG-5833 ]]

Test Plan: Made sure that when new members were added, they are still assigned the default_role. Also ran the query before and after in TablePlus and made sure that the old default_role column value was equal to the new id column value.

Reviewers: atul, ginsu, tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D9979