[Flow202][native][skip-ci] [37/x] Allow removeEditMode to take a TabAction
Summary:
We call `removeEditMode` in `native/navigation/tab-router.js`, so it needs to be able to support a `TabAction` as well as `ChatRouterNavigationAction`. Separating this diff out because the changes are non-obvious.

NOTE: CI will fail on this diff. I considered the possibility of fixing Flow errors BEFORE upgrading Flow, but it wasn't possible... in some cases, the fixes to support the new version of Flow caused errors in the old version. I could have hidden these type errors with `$FlowFixMe` lines and then later revert those, but that seemed like too much busy work.

Depends on D10020

Test Plan: Confirm the Flow errors go away

Reviewers: tomek

Differential Revision: https://phab.comm.dev/D10021