[keyserver] Update usage of threads.default_role in thread-fetchers.js
Summary:
We need to get rid of usages of `threads.default_role` throughout the codebase so we can eventually drop the column.

This diff handles the usage in `thread-fetchers.js`

Existing code: Fetches the default_role when constructed threadInfos to set the isDefault flag for each role in threadInfo.roles

Depends on D9979

Addresses part of [[ https://linear.app/comm/issue/ENG-5833/update-code-that-uses-default-role-from-mariadb | ENG-5833 ]]

Test Plan:
Updated my query and logged the `threadInfo.roles`. Confirmed that for each role, `isDefault` was set to true for all `Members` and false otherwise.

Another test I did was to compare the logged `threadInfo.roles` both before and after the query change and they were identical

Reviewers: atul, ginsu, tomek

Reviewed By: tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D9980