[keyserver] Update usage of threads.default_role in thread-fetchers.js
Summary:
We need to get rid of usages of `threads.default_role` throughout the codebase so we can eventually drop the column.

This diff handles the usage in `thread-fetchers.js`

Depends on D9979

Addresses part of [[ https://linear.app/comm/issue/ENG-5833/update-code-that-uses-default-role-from-mariadb | ENG-5833 ]]

Test Plan: Updated my query and logged the threadInfo.roles. Confirmed that for each role, isDefault was set to true for all Members and false otherwise. An alternate test is to compare the logged threadInfo.roles both before and after the query change since they should be identical (confirmed this is true).

Reviewers: atul, ginsu, tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D9980