Extend showNotification method to cover the case of decryption error notitication
Summary:
This differential refactors `showNotification` method in native desktop code so that it can handle future case of notification that we failed to decrypt. If
we fail to decrypt notification we will either display generic "Comm notification" or add error stack trace to the body depending on whether current user is a staff
member or we are running DEV environment.

Test Plan:
1.  Modify `listenForNotifications` method so that on each notification it calls `showNotification` twice. The first time with real notification and teh
second time with notification that matched decryption error notification structure (`error` and `displayErrorMessage` fields)
2. Check navigation behaviour of error message notification.

Reviewers: kamil, michal

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D9963