[keyserver] Fetch invite link blob holder
Summary:
Update the logic where we fetch invite links to also get the holder. We don't need to show this to a client and we're filtering this field out for them.

Depends on D9966

Test Plan: Manually add a holder to a link and check if it is fetched correctly.

Reviewers: michal, bartek, rohan