[keyserver] Update usage of threads.default_role in thread-creator.js
Summary:
We need to get rid of usages of `threads.default_role` throughout the codebase so we can eventually drop the column.

This diff handles the usage in `thread-creator.js`

Existing code: When creating a thread, the roles from createInitialRolesForNewThread are used to populate the role ID into threads.default_role for that specific row. We should no longer populate this field.

Depends on D9980

Addresses part of [[ https://linear.app/comm/issue/ENG-5833/update-code-that-uses-default-role-from-mariadb | ENG-5833 ]]

Test Plan:
1. Ran `ALTER TABLE threads MODIFY COLUMN default_role bigint(20) NULL;` to allow NULL values in the column for now (until the column is dropped)
2. Created a new thread and checked to make sure that no role got inserted into `default_role`

Reviewers: atul, ginsu, tomek

Reviewed By: tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D9981