[native] Increase the number of threads available for rust
Summary:
[ENG-5561](https://linear.app/comm/issue/ENG-5561)

We don't want to block other tasks on backup operations. We can increase the number of used thread in tokio runtime to the default (the number of cores of the system) so other operations can still proceed. This shouldn't cause a problem for NSE (there were worries that too many threads could make the notifications less likely to be delivered) because `rust_native_library` isn't even compiled for the notification service.

Test Plan:
Test that the app runs on physical iOS and Android devices

Reviewers: marcin, kamil

Subscribers: