[Flow202][native][skip-ci] [42/x] Move ImageSource types to native/types/react-native.js and use it to parameterize ref
Summary:
When we need to import a type from a file inside of the `react-native` package, we have a practice of only specifying the path to that file in one place in our codebase. The internals are subject to change, so it's easier to handle upgrades this way.

This diff dedups the `ImageSource` imports, and also adds a new usage of `ImageSource` to parameterize a `useRef` in `EncryptedImage`.

Separating this diff out since it affects multiple files.

NOTE: CI will fail on this diff. I considered the possibility of fixing Flow errors BEFORE upgrading Flow, but it wasn't possible... in some cases, the fixes to support the new version of Flow caused errors in the old version. I could have hidden these type errors with `$FlowFixMe` lines and then later revert those, but that seemed like too much busy work.

Depends on D10025

Test Plan: Confirm the Flow errors go away

Reviewers: bartek

Subscribers: