[Flow202][native][skip-ci] [58/x] Destructure this.lastMessageSent before checking its value
Summary:
The new version of Flow appears to be worried that the value of `this.lastMessageSent` will change after it's checked. The error goes away if I assign the value to a local variable before checking it.

NOTE: CI will fail on this diff. I considered the possibility of fixing Flow errors BEFORE upgrading Flow, but it wasn't possible... in some cases, the fixes to support the new version of Flow caused errors in the old version. I could have hidden these type errors with `$FlowFixMe` lines and then later revert those, but that seemed like too much busy work.

Depends on D10041

Test Plan: Confirm the Flow errors go away

Reviewers: tomek

Subscribers: