[identity] Make authenticated X3DH RPCs use userID
Summary:

Main part of [[ https://linear.app/comm/issue/ENG-5659/use-user-id-for-lookup-instead-of-usernamewallet-address-in-x3dh | ENG-5659 ]]. This diff makes it happen with minimal changes to the existing code.

Depends on D9957

Test Plan:

Called the RPC with a userID and it worked as expected. Compared the results with the old username RPCs in unauth proto.

Reviewers: varun!

Subscribers: