[grpc clients] Update Rust lib to use authenticated RPCs
Summary:
Updated the client method to use the auth proto RPCs for GetOutboundKeys and GetInboundKeys.
Needed to add more params, cuz we need two distinct userID/deviceID pairs: for current user (auth) and for the user we want the keys for.

Depends on D10048.

Test Plan: Native and keyserver compiles. These functions aren't called anywhere at the moment in the JS code

Reviewers: varun

Reviewed By: varun

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D10049