[web] use default `updatesCurrentAsOf` in transform if value is missing
Summary:
This is just an enhancement, I tested and this not cause any issues, but I prefer to add this for consistency.

In D9948 persistence to this field was enabled, but on the first rehydration this value could be undefined.

Alternative solution is adding migration, not sure what is better here.

Test Plan:
1. Reverting D9948
2. Use web app
3. Revert commit reverting D9948
4. Check if `updatesCurrentAsOf` is not rehydrated as undefined.

Reviewers: inka, tomek, michal

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D10069