[lib] Re-introduce `RelativeMemberInfo` (`[Legacy/MinimallyEncoded]RelativeMemberInfo`)
Summary:
Similar to D10068 and D10072, but for `RelativeMemberInfo`.

Note: There are a few places that we haven't updated to support `MinimallyEncodedRelativeMemberInfo` yet (specifically `getRelativeMemberInfos`). Those will be handled later in the stack (and will be easier to identify because we'll be able to search for `LegacyRelativeMemberInfo` and find all instances.

---

Depends on D10073

Test Plan: `flow`/CI/etc

Reviewers: ashoat, ginsu, tomek, rohan

Differential Revision: https://phab.comm.dev/D10074