[lib/native/web] lift customServer redux state into lib
Summary:
At the moment, the `customServer` redux field lives only in native. Since we want to use this state in web as well, we will need to lift the redux logic into lib. This diff handles lifting the redux state into lib

Depends on D10090

Test Plan:
`flow` and confirmed that the custom server on native works as expected

Here is a demo video also showing that the custom server field also works on `web` too
{F911854}

Reviewers: inka, michal, rohan

Reviewed By: inka

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D10092