[lib/native/web] lift customServer redux state into lib
Summary: At the moment, the `customServer` redux field lives only in native. Since we want to use this state in web as well, we will need to lift the redux logic into lib. This diff handles lifting the redux state into lib

Test Plan: `flow` and confirmed that the custom server on native works as expected

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D10092