[web] Replace {||} with {} in identity-structs.cjs.flow
Summary:
Following the Flow upgrade, I was experimenting with the various [codemods in](https://www.npmjs.com/package/flow-upgrade) `flow-upgrade`. I gave `removeExplicitlyExactObjectTypeSyntax` a try, since we have `exact_by_default=true` set in our `.flowconfig`s, and as such `{}` is an alias to `{||}`.

The codemod found a couple locations in this file. For the sake of consistency, I replaced them with `{}`, like we do elsewhere in our codebase.

Depends on D10087

Test Plan: Flow

Reviewers: varun, bartek

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D10088