[web] save the recently deleted server url to customServer redux state
Summary:
if a server is removed from the list, and the user is staff or in dev mode, we want this url to be saved to customServer. This diff implements this functionality.

Linear task: https://linear.app/comm/issue/ENG-5498/introduce-onpress-remove-selected-keyserver-function

Depends on D10092

Test Plan:
Please see demo video below

{F901210}

Reviewers: inka, rohan, michal

Reviewed By: rohan

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D10093