[web] implement onClickAddKeyserver callback
Summary: When a user clicks on the add keyserver button their keyserver should be stored in the keyserver store. `onClickAddKeyserver` handles this funtionality. Additionally, if the user is a staff or in dev mode, we want the input field to be pre-populated with the value of `customServer`

Test Plan: Please see the demo video below

Reviewers:

Subscribers: