[web] implement onClickAddKeyserver callback
Summary:
When a user clicks on the add keyserver button their keyserver should be stored in the keyserver store. `onClickAddKeyserver` handles this functionality. Additionally, if the user is a staff or in dev mode, we want the input field to be pre-populated with the value of `customServer`

Linear task: https://linear.app/comm/issue/ENG-4726/dispatch-add-keyserver-redux-action-when-onpress-save-button-is

Depends on D10093

Test Plan:
Please see the demo video below

{F901280}

Reviewers: inka, rohan, michal

Reviewed By: inka

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D10094