[native] Use consistent panel colors in QR code screens
Summary:
All of the components in `qr-code-screen.react.js` use a color from the panel color family, but the background in `qr-code-sign-in-navigator.js` use a color from the modal family. This diff just swaps the color out to instead remain consistent and use a color from the panel family.

This should have no impact visually because `panelBackground` is the same as `modalBackground` (`designSystemColors.shadesBlack95`)

Test Plan: Confirmed that there is no visual change

Reviewers: ginsu, atul

Reviewed By: ginsu

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D10100