[lib] Add connectionIssue field
Summary:
issue: https://linear.app/comm/issue/ENG-5990/add-setconnectionissue-action
In this stack I will be replacing logouts dispatched from the socket with seting this new `connectionIssue` filed. This is because we don't want to log the user out of the app if we can't connect to one keyserver. We also don't want to log the user out of this keyserver, because that would result in the client removing the keyserver from the store. Instead, we decided to store information about the problem that the client is having with communicating with the keyserver.
See the discussion in [[ https://linear.app/comm/issue/ENG-5808/refactor-logout-action-usages-connection-issue#comment-710a3e90 | ENG-5808 ]] for more detials.

Test Plan: flow

Reviewers: michal, kamil, ginsu

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D10287