[identity] increase number of tasks and enable stickiness on load balancer
Summary:
Resolves https://linear.app/comm/issue/ENG-5909/increase-number-of-identity-service-nodes-on-stagingprod

We want to get used to thinking of the identity service as a distributed system. Enabling session stickiness lets us send _finish RPC requests to the same task as the corresponding _start requests by setting a cookie.

Test Plan:
https://gist.github.com/vdhanan/07cd60c9474f76d63fa10ef9d909b8ef

used the lb cookie to send a bunch of requests in quick succession, saw in the staging logs that they went to the same task until the cookie expired, at which point they went to both tasks

Reviewers: bartek!

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D10102