[lib] Remove KeyserverInfoPartial type
Summary:
Following the Flow upgrade, I'm looking at replacing all usages of `$Shape` in our codebase. Context [here](https://medium.com/flow-type/announcing-partial-required-flow-utility-types-catch-annotations-3a32f0bf2a20).

When I swapped the `$Shape` in `KeyserverInfoPartial` into a `Partial`, it revealed a type error. The issue is that the `urlPrefix` and `connectionStatus` we pass to `createBoundServerCallsSelector` are nullable, but that function requires them to be set.

I investigated further and found that we don't actually use `KeyserverInfoPartial` in any case where its "partialness" is necessary... it appears we just get the `KeyserverInfos` from the Redux store and pass them in. As such, I deleted the type and replaced usages with `KeyserverInfos`.

Depends on D10082

Test Plan: Flow

Reviewers: inka, michal

Reviewed By: inka

Subscribers: ginsu, tomek

Differential Revision: https://phab.comm.dev/D10083