[native] Update update-roles-and-permissions-test-data.js to include specialRole field
Summary: There are some places in the codebase that we hardcode `isDefault`. Here, we'll want to introduce the `specialRole` field alongside, so when we deprecate `isDefault`, no additional changes will be needed. Some are important (like one instance in `thread-utils.js` and the instance in `role-creator.js` where we set up initial roles for threads), and others are just test data. Nonetheless, I thought I may as well update all of them for consistency.

Part of [[ https://linear.app/comm/issue/ENG-5994/introduce-specialrole-field-in-roleinfo | ENG-5993 ]]

This diff handles `update-roles-and-permissions-test-data.js`

Depends on D10114

Test Plan: TODO

Reviewers: atul, ginsu, tomek

Subscribers: