[keyserver] Create a blob while adding / editing an invite link
Summary:
Every time we're creating / updating a link, we need to make sure that the blob is created.
We start with blob creation because it is a shared place between keyservers and links should be globally unique - a keyserver should first check if a blob can be created which means that other keyservers haven't used that link.

Depends on D10131

Test Plan:
Create a link and check if a correct blob is created.
Update a link and check if a blob is created and an old blob holder is deleted.

Reviewers: michal, bartek, rohan

Reviewed By: rohan

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D10134