[lib] Get rid of switchMemberAdminRoleInThread
Summary:
Working on enabling the `unused-promise` Flow lint ([ENG-5907](https://linear.app/comm/issue/ENG-5907/set-up-unused-promise-flow-lint-for-keyserver)), and this got flagged because of the `dispatchActionPromise` usage.

I looked but I couldn't see any usages of this function in the codebase. I think the last usage was removed in D8155.

Depends on D10158

Test Plan: Flow

Reviewers: rohan

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D10159