[lib] Don't return a promise from ignorePromiseRejections
Summary:
This diff is a no-op since no callsite actually pays attention to the result here.

The reason we need this is to avoid having to prefix every callsite with `void`. Otherwise, the `unused-promise` Flow lint that gets enabled in a following diff would fire at each callsite.

This diff avoids that by prefixing with `void` inside the definition of `ignorePromiseRejections`.

Depends on D10159

Test Plan: Flow

Reviewers: michal

Reviewed By: michal

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D10160