[lib] Wrap spinPromises in ignorePromiseRejections
Summary:
The `unused-promise` Flow lint revealed an issue here. We use `SequentialPromiseResolver` in `keyserver`, where an unhandled promise rejection could crash the whole server. But we aren't doing anything to handle a potential promise rejection here.

This diff wraps `spinPromises` in `ignorePromiseRejections`. This way, if the promise rejects we will print an error instead of crashing the whole keyserver.

Depends on D10160

Test Plan: Flow

Reviewers: michal

Reviewed By: michal

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D10161