[web] Hide that InputState.sendTextMessage returns a Promise
Summary:
While working on the `unused-promise` Flow lint, I noticed that `sendTextMessage` doesn't actually `await` the call to `sendTextMessageAction`. This makes it dangerous to expose as an API to callers, who probably assume the keyserver call is complete when the `Promise` resolves.

I could make it `await` here, but since apparently no callsites need a `Promise` currently, I decided to just make the API return a `mixed` here for now.

Depends on D10161

Test Plan: Flow

Reviewers: atul, bartek

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D10162