[keyserver] set default identity socket addr based on NODE_ENV value
Summary: if NODE_ENV is "development", we try to connect to staging. if it's anything else, we try to connect to the prod identity service.

before landing this, i will let devs know that in order to talk to their local identity service, they'll now need to have a well-formed secrets/identity_service_config.json file or set the equivalent env var

Test Plan: ran `yarn dev` and connected to the staging instance

Reviewers: bartek!, ashoat

Subscribers: