[service] Move services lib to shared
Summary: Part of [ENG-5555](https://linear.app/comm/issue/ENG-5555) -> in order to move the backup client from commtest to shared (to use it from native) we need to move the `comm_services_lib` to shared, because the client uses e.g. the `UserIdentity` struct. It would be good to have a common library for all Rust code in our codebase anyway. This diff also renames the library to `comm_lib` as it will stop being used only for services.

Test Plan: Find-and-replace `comm_services_lib` with `comm_lib`. Test if all services build.

Reviewers: bartek, varun, kamil

Reviewed By: bartek

Subscribers: wyilio, ashoat, tomek

Differential Revision: https://phab.comm.dev/D10256