[services] Fix warnings
Summary:
Fixes two small warnings:
- unused import
- deprecation of `from_utc` -> replaced with `from_naive_utc_and_offset`

The second one is a bit tricker because the deprecation notice was introduced in a patch version, and our services used two different versions of the library. The newer version had the deprecation notice. But we can't just update all the callsites because the older version doesn't have the replacement function. Because of that I also pinned this library in all of our dependencies to the newer patch version.

Depends on D10256

Test Plan: Check if services compile.

Reviewers: bartek, varun, kamil

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D10257