[keyserver] Fetch all the links when a viewer is a script viewer
Summary:
Currently we're fetching only the links for communities where a viewer is a member. For the migration to work we need to change that and fetch all the links in a script.

Depends on D10134

Test Plan:
Check if in the web app we're receiving only the links for communities where a viewer is a member.
Check if using script viewer changes the behavior and all the links are fetched.

Reviewers: michal, bartek, rohan

Subscribers: