[keyserver] Introduce a function that allows fetching all the primary links
Summary:
Currently we're fetching only the links for communities where a viewer is a member. For the migration to work we need to be able to fetch all the links - this diff introduces a function that allows it.

Depends on D10134

Test Plan:
Check if in the web app we're receiving only the links for communities where a viewer is a member.
Check if calling `fetchAllPrimaryInviteLinks` returns all the links.

Reviewers: michal, bartek, rohan

Reviewed By: rohan

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D10150